+++ NEWS.md
@@ -3,8 +3,7 @@
 - use more accurate data.summary and path.data object names.
 - amend drawPath() syntax.
 - amend milePosts().
-- update zero-length arrow diagnostics.
-- amend zero-length arrow fix (use "|" in place of "<").
+- use shape::Arrowhead() for mileposts (zero-length arrow fix).
 
 
 # cholera 0.9.1.9080